Config File Changes TOPICS: Config file, System Class, Results directory
The config file now includes user input requirements for the location of
the folder containing the raw concentration curve data. It also requires
the specification of a working directory "wd" that the results and,
eventually, compiled config file will be saved to.

All subsequent file changes are to correct previous code base behavior
based on the old config file format. I changed the E_curves.py script to
not include a main function as it is now performed with every run of
model_eval.py. Next commit should probably move E_curves.py to the
modules folder as it is no longer ran as main in the current code base.

Model eval was changed accordingly to include the generation of the
E_curves and Etheta_curves for each run of a model.