SOBOL SENSTIVITY ANALYSIS and SETUP.py
In sensitivity_analysis.ipynb, I finished my
sobol global sensitivity analysis. For this, I
had to install a package called SALib which I
added to setup.py. Even though it's not used
anywhere in src right now, it would be cool to
add a sensitivity analysis featuer in the
code base someday.